Move the logging flag to pkg/logs/flag for reusability across the commands
