+++ cmd/ignite/cmd/cmdutil/flags.go
@@ -4,7 +4,6 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/sirupsen/logrus"
 	"github.com/spf13/pflag"
 	api "github.com/weaveworks/ignite/pkg/apis/ignite"
 	meta "github.com/weaveworks/ignite/pkg/apis/meta/v1alpha1"
@@ -85,33 +84,6 @@ func OCIImageRefVar(fs *pflag.FlagSet, ptr *meta.OCIImageRef, name, usage string
 
 func OCIImageRefVarP(fs *pflag.FlagSet, ptr *meta.OCIImageRef, name, shorthand, usage string) {
 	fs.VarP(&OCIImageRefFlag{value: ptr}, name, shorthand, usage)
-}
-
-type LogLevelFlag struct {
-	value *logrus.Level
-}
-
-func (lf *LogLevelFlag) Set(val string) error {
-	var err error
-	*lf.value, err = logrus.ParseLevel(val)
-	return err
-}
-
-func (lf *LogLevelFlag) String() string {
-	if lf.value == nil {
-		return ""
-	}
-	return lf.value.String()
-}
-
-func (lf *LogLevelFlag) Type() string {
-	return "loglevel"
-}
-
-var _ pflag.Value = &LogLevelFlag{}
-
-func LogLevelFlagVar(fs *pflag.FlagSet, ptr *logrus.Level) {
-	fs.Var(&LogLevelFlag{value: ptr}, "log-level", "Specify the loglevel for the program")
 }
 
 type NetworkModeFlag struct {
